+++ src/compiler/binder/mod.rs
@@ -2,3 +2,5 @@ mod lines_0_500;
 pub use lines_0_500::*;
 mod lines_500_1000;
 pub use lines_500_1000::*;
+mod lines_1000_2000;
+pub use lines_1000_2000::*;
